fixtop & fixbottom of form-child move along
with parent's y coord to have correct fixtop & fixbottom. These are used
as reference for the topmost and bottommost of a form and are useful eg
when we need the height of the staff in calculations. fixheight is
computed only at init-time and never changes. In all these values is the
value of Y coord (calced at canvas init-time) hence y-offset & possibly
absy included.